refactoring and implementation for export C matrix
