+++ src/ansys/aedt/core/application/analysis.py
@@ -57,7 +57,6 @@ from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
 from ansys.aedt.core.generic.settings import settings
 from ansys.aedt.core.modules.boundary.layout_boundary import NativeComponentObject
 from ansys.aedt.core.modules.boundary.layout_boundary import NativeComponentPCB
-from ansys.aedt.core.modules.boundary.maxwell_boundary import MaxwellParameters
 from ansys.aedt.core.modules.design_xploration import OptimizationSetups
 from ansys.aedt.core.modules.design_xploration import ParametricSetups
 from ansys.aedt.core.modules.solve_setup import Setup
@@ -2314,113 +2313,6 @@ class Analysis(Design, object):
         from ansys.aedt.core.generic.general_methods import _dim_arg
 
         return _dim_arg(value, units)
-
-    @pyaedt_function_handler(file_path="output_file", setup_name="setup")
-    def export_rl_matrix(
-        self,
-        matrix_name,
-        output_file,
-        is_format_default=True,
-        width=8,
-        precision=2,
-        is_exponential=False,
-        setup=None,
-        default_adaptive=None,
-        is_post_processed=False,
-    ):
-        """Export R/L matrix after solving.
-
-        Parameters
-        ----------
-        matrix_name : str
-            Matrix name to be exported.
-        output_file : str
-            Output file path to export R/L matrix file to.
-        is_format_default : bool, optional
-            Whether the exported format is default or not.
-            If False the custom format is set (no exponential).
-        width : int, optional
-            Column width in exported .txt file.
-        precision : int, optional
-            Decimal precision number in exported \\*.txt file.
-        is_exponential : bool, optional
-            Whether the format number is exponential or not.
-        setup : str, optional
-            Name of the setup.
-        default_adaptive : str, optional
-            Adaptive type.
-        is_post_processed : bool, optional
-            Boolean to check if it is post processed. Default value is ``False``.
-
-        Returns
-        -------
-        bool
-            ``True`` when successful, ``False`` when failed.
-        """
-        if not self.solution_type == "EddyCurrent":
-            self.logger.error("RL Matrix can only be exported if solution type is Eddy Current.")
-            return False
-        matrix_list = [bound for bound in self.boundaries if isinstance(bound, MaxwellParameters)]
-        if matrix_name is None:
-            self.logger.error("Matrix name to be exported must be provided.")
-            return False
-        if matrix_list:
-            if not [
-                matrix
-                for matrix in matrix_list
-                if matrix.name == matrix_name or [x for x in matrix.available_properties if matrix_name in x]
-            ]:
-                self.logger.error("Matrix name doesn't exist, provide and existing matrix name.")
-                return False
-        else:
-            self.logger.error("Matrix list parameters is empty, can't export a valid matrix.")
-            return False
-
-        if output_file is None:
-            self.logger.error("File path to export R/L matrix must be provided.")
-            return False
-        elif os.path.splitext(output_file)[1] != ".txt":
-            self.logger.error("File extension must be .txt")
-            return False
-
-        if setup is None:
-            setup = self.active_setup
-        if default_adaptive is None:
-            default_adaptive = self.design_solutions.default_adaptive
-        analysis_setup = setup + " : " + default_adaptive
-
-        if not self.available_variations.nominal_w_values_dict:
-            variations = ""
-        else:
-            variations = " ".join(
-                f"{key}=\\'{value}\\'" for key, value in self.available_variations.nominal_w_values_dict.items()
-            )
-
-        if not is_format_default:
-            try:
-                self.oanalysis.ExportSolnData(
-                    analysis_setup,
-                    matrix_name,
-                    is_post_processed,
-                    variations,
-                    output_file,
-                    -1,
-                    is_format_default,
-                    width,
-                    precision,
-                    is_exponential,
-                )
-            except Exception:
-                self.logger.error("Solutions are empty. Solve before exporting.")
-                return False
-        else:
-            try:
-                self.oanalysis.ExportSolnData(analysis_setup, matrix_name, is_post_processed, variations, output_file)
-            except Exception:
-                self.logger.error("Solutions are empty. Solve before exporting.")
-                return False
-
-        return True
 
     @pyaedt_function_handler()
     def change_property(self, aedt_object, tab_name, property_object, property_name, property_value):
